[수정] 통계 관련 비즈니스 로직 수정
전체 갯수를 keySet이 아닌 findAll()한 리스트를 통해 가져오도록 수정
KeySet의 전체 갯수는 중복을 제외한 수임

+++ src/main/java/com/cloth/wardrobe/dto/statistics/ContentForStatistics.java
@@ -9,6 +9,6 @@ import lombok.Data;
 public class ContentForStatistics {
     String name;
     StatisticsType type;
-    int percent;
+    double percent;
     int count;
 }
